# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KVRecord": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "description": "demographics | condition | medication | allergy | lab | encounter | icd10 | drug",
            "title": "Kind",
            "type": "string"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "source_id": {
            "readOnly": true,
            "title": "Source Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "kind",
          "title",
          "data",
          "source_id"
        ],
        "title": "KVRecord",
        "type": "object"
      },
      "LabWrite": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "lab": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Lab",
            "type": "string"
          },
          "unit": {
            "maxLength": 20,
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "lab",
          "value",
          "unit",
          "date"
        ],
        "title": "LabWrite",
        "type": "object"
      },
...
      "Origin": {
        "enum": [
          "kv",
          "vector"
        ],
        "title": "Origin",
        "type": "string"
      },
      "PatientRecord": {
        "properties": {
          "notes": {
            "description": "Note metadata: doc_id, title, date.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Notes",
            "type": "array"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/KVRecord"
            },
            "title": "Records",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/PatientSummary"
          }
        },
        "required": [
          "summary",
          "records",
          "notes"
        ],
        "title": "PatientRecord",
        "type": "object"
      },
      "PatientSummary": {
        "properties": {
          "age": {
            "title": "Age",
            "type": "integer"
          },
          "birth_date": {
            "format": "date",
            "title": "Birth Date",
            "type": "string"
          },
          "conditions": {
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "lab_count": {
            "title": "Lab Count",
            "type": "integer"
          },
          "last_encounter": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Encounter"
          },
          "medication_count": {
            "title": "Medication Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note_count": {
            "title": "Note Count",
            "type": "integer"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "sex": {
            "title": "Sex",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "name",
          "sex",
          "birth_date",
          "age",
          "conditions",
          "medication_count",
          "lab_count",
          "note_count"
        ],
        "title": "PatientSummary",
        "type": "object"
      },
...
      "SourceView": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "origin": {
            "$ref": "#/components/schemas/Origin"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "origin",
          "title",
          "text"
        ],
        "title": "SourceView",
        "type": "object"
      },
...
      "WriteResult": {
        "properties": {
          "data_version": {
            "title": "Data Version",
            "type": "integer"
          },
          "invalidated_cache_entries": {
            "title": "Invalidated Cache Entries",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key",
          "invalidated_cache_entries",
          "data_version"
        ],
        "title": "WriteResult",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Hybrid KV + vector retrieval for a clinical knowledge assistant. **Educational prototype. Not medical advice.**",
    "title": "MedMemory API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit": {
      "get": {
        "operationId": "audit_tail_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "n",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Audit Tail V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Tail",
        "tags": [
          "ops"
        ]
      }
    },
...
    "/v1/kv": {
      "get": {
        "operationId": "kv_scan_v1_kv_get",
        "parameters": [
          {
            "in": "query",
            "name": "prefix",
            "required": true,
            "schema": {
              "minLength": 3,
              "title": "Prefix",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KVRecord"
                  },
                  "title": "Response Kv Scan V1 Kv Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kv Scan",
        "tags": [
          "records"
        ]
      }
    },
    "/v1/kv/{key}": {
      "get": {
        "operationId": "kv_get_v1_kv__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KVRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kv Get",
        "tags": [
          "records"
        ]
      }
    },
...
    "/v1/patients": {
      "get": {
        "operationId": "patients_v1_patients_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PatientSummary"
                  },
                  "title": "Response Patients V1 Patients Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patients",
        "tags": [
          "records"
        ]
      }
    },
    "/v1/patients/{pid}": {
      "get": {
        "operationId": "patient_v1_patients__pid__get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patient",
        "tags": [
          "records"
        ]
      }
    },
    "/v1/patients/{pid}/labs": {
      "post": {
        "description": "Write path: KV put, then write-through invalidation of every cached answer tagged\nwith this patient (and the data version bump orphans any stragglers).",
        "operationId": "add_lab_v1_patients__pid__labs_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Lab",
        "tags": [
          "records"
        ]
      }
    },
    "/v1/query": {
      "post": {
        "operationId": "query_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query",
        "tags": [
          "query"
        ]
      }
    },
    "/v1/query/stream": {
      "post": {
        "operationId": "query_stream_v1_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "SSE: meta, retrieval, token\u2026, final"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Stream",
        "tags": [
          "query"
        ]
      }
    },
...
    "/v1/search": {
      "post": {
        "operationId": "search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "vector"
        ]
      }
    },
    "/v1/sources/{source_id}": {
      "get": {
        "operationId": "source_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source",
        "tags": [
          "vector"
        ]
      }
    }
  }
}